+++ apps/block_scout_web/lib/block_scout_web/notifier.ex
@@ -229,6 +229,10 @@ defmodule BlockScoutWeb.Notifier do
     Endpoint.broadcast("addresses:#{to_string(address_hash)}", "smart_contract_was_verified", %{})
   end
 
+  def handle_event({:chain_event, :optimism_deposits, :realtime, deposits}) do
+    broadcast_optimism_deposits(deposits, "optimism_deposits:new_deposits", "deposits")
+  end
+
   def handle_event(_), do: nil
 
   def fetch_compiler_version(compiler) do
@@ -401,6 +405,10 @@ defmodule BlockScoutWeb.Notifier do
         transaction: transaction
       })
     end
+  end
+
+  defp broadcast_optimism_deposits(deposits, deposit_channel, event) do
+    Endpoint.broadcast(deposit_channel, event, %{deposits: deposits})
   end
 
   defp broadcast_token_transfer(token_transfer) do