+++ apps/explorer/priv/repo/migrations/20230216135703_add_op_transaction_batches_table.exs
@@ -1,0 +1,14 @@
+defmodule Explorer.Repo.Migrations.AddOpTransactionBatchesTable do
+  use Ecto.Migration
+
+  def change do
+    create table(:op_transaction_batches, primary_key: false) do
+      add(:l2_block_number, :bigint, null: false, primary_key: true)
+      add(:epoch_number, :bigint, null: false)
+      add(:l1_tx_hashes, {:array, :bytea}, null: false)
+      add(:l1_tx_timestamp, :"timestamp without time zone", null: false)
+
+      timestamps(null: false, type: :utc_datetime_usec)
+    end
+  end
+end